requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
dj-database-url==3.1.2
Django==5.2.12
django-allauth==65.15.0
django-cors-headers==4.9.0
djangorestframework==3.17.0
gunicorn==25.1.0
idna==3.11
oauthlib==3.3.1
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0

redis==5.0.1
django-redis==5.4.0
yfinance
pandas==2.2.3
openpyxl==3.1.5
